requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
pydantic==2.11.1
pydantic-settings==2.8.1
python-multipart==0.0.20
PyJWT==2.10.1
firebase-admin==6.7.0
aiofiles==24.1.0
email-validator==2.2.0
httpx==0.28.1
passlib==1.7.4
pytest==8.3.5
